fix de cards en pagina de habitaciones
