+++ src/App.jsx
@@ -6,31 +6,9 @@ import ReservationForm from "./components/pages/ReservationForm";
 import Login from "./components/pages/Login/Login.jsx";
 import Gallery from "./components/pages/Gallery.jsx";
 import FilterRoomsContain from "./components/FilterRoomsContainer/FilterRoomsContainer.jsx";
-import useFetch from "./hooks/useFetch.jsx";
-import CardRoom from "./components/common/CardRoom/CardRoom.jsx";
-import { Container } from "react-bootstrap";
-
-const RoomsContainer = () => {
-  const { data, loading, error } = useFetch({ endPoint: `/rooms` });
-
-  return (
-    <Container>
-      {data !== null &&
-        data.length > 0 &&
-        data.map((room) => (
-          <CardRoom
-            key={room.id}
-            room_name={room.room_name}
-            price={room.price}
-            image={room.image}
-            broad_description={room.broad_description}
-            benefits={room.benefits}
-          />
-        ))}
-    </Container>
-  );
-};
 import NavbarComponent from "./components/common/Navbar/NavbarComponent.jsx";
+import RoomsContainer from "./components/pages/Rooms.jsx";
+
 
 function App() {
   return (
@@ -56,7 +34,7 @@ const AppContent = () => {
             element={<FilterRoomsContain />}
           />
         </Route>
-        <Route path="/habitaciones" element={<RoomsContainer />} />
+        <Route path="/habitaciones" element={<RoomsContainer/>} />
         <Route path="/sobre-nosotros" element={<QuienesSomos />} />
         <Route path="/admin" element={<h1>Administrador</h1>} />
         <Route path="/galeria" element={<Gallery />} />
